# reference/public-v1.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "{{siteName}} Public API",
    "version": "2.0.0",
    "description": "Anonymous, read-only API for clients and Agents. No API key is required.\n\nPicking an endpoint: show recent or today's items -> /api/v1/items. Show what is trending right now -> /api/v1/hot-topics. Show the daily digest -> /api/v1/dailies/latest. Keep a complete local copy of the selected set -> /api/v1/selected/snapshot once, then /api/v1/selected/changes forever.\n\nHow often to poll: use each response's Cache-Control s-maxage as the minimum interval, and always send If-None-Match: an unchanged resource answers 304 with an empty body. On 429 or 503, honour Retry-After. There is no push channel (no SSE, webhooks or streaming): poll on an interval.\n\nTitles and summaries are generated from third-party sources; the original is always linked in links.original.",
    "termsOfService": "{{siteUrl}}/terms"
  },
  "servers": [
    {
      "url": "{{siteUrl}}"
    }
  ],
...
  "paths": {
...
    "/api/v1/dailies": {
      "get": {
        "operationId": "dailies",
        "summary": "List daily-report archive entries",
        "description": "Date index of past daily reports, newest first. Reports are published once a day at 08:00 Asia/Shanghai, so checking this more than a few times a day is wasted work.",
        "parameters": [
          {
            "$ref": "#/components/parameters/ActorUserAgent"
          },
          {
            "$ref": "#/components/parameters/IfNoneMatch"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 180,
              "default": 30
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailiesResponse"
                }
              }
            }
          },
          "304": {
            "description": "Not Modified",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Machine-readable problem response.",
            "headers": {
              "X-Request-Id": {
                "description": "Support correlation identifier for this failed request.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Machine-readable problem response.",
            "headers": {
              "X-Request-Id": {
                "description": "Support correlation identifier for this failed request.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "405": {
            "description": "Machine-readable problem response.",
            "headers": {
              "X-Request-Id": {
                "description": "Support correlation identifier for this failed request.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Machine-readable problem response.",
            "headers": {
              "X-Request-Id": {
                "description": "Support correlation identifier for this failed request.",
                "schema": {
                  "type": "string"
                }
              },
              "Retry-After": {
                "description": "Minimum number of seconds to wait before retrying.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Machine-readable problem response.",
            "headers": {
              "X-Request-Id": {
                "description": "Support correlation identifier for this failed request.",
                "schema": {
                  "type": "string"
                }
              },
              "Retry-After": {
                "description": "Minimum number of seconds to wait before retrying.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "ActorUserAgent": {
        "name": "User-Agent",
        "in": "header",
        "required": false,
        "description": "The client's normal User-Agent (optional; any value works).",
        "schema": {
          "type": "string",
          "maxLength": 2048
        },
        "example": "my-reader/1.0"
      },
      "IfNoneMatch": {
        "name": "If-None-Match",
        "in": "header",
        "required": false,
        "description": "Previously received ETag for this exact URL. A match returns 304 without a response body.",
        "schema": {
          "type": "string"
        }
      }
    },
    "schemas": {
...
      "Problem": {
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code",
          "requestId"
        ],
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          }
        }
      },
      "Attribution": {
        "type": "object",
        "description": "Machine-readable {{siteName}} provenance for traceability. Retain it where technically feasible. Attribution alone is not authorization; see the public terms.",
        "required": [
          "name",
          "url"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "url": {
            "type": "string",
            "format": "uri"
          }
        }
      },
...
      "DailiesResponse": {
        "type": "object",
        "required": [
          "schemaVersion",
          "count",
          "items"
        ],
        "properties": {
          "schemaVersion": {
            "const": 1
          },
          "count": {
            "type": "integer",
            "minimum": 0
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DailyEntry"
            }
          }
        }
      },
      "DailyEntry": {
        "type": "object",
        "required": [
          "date",
          "generatedAt",
          "leadTitle",
          "leadParagraph",
          "links"
        ],
        "properties": {
          "date": {
            "type": "string",
            "format": "date"
          },
          "generatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "leadTitle": {
            "type": [
              "string",
              "null"
            ]
          },
          "leadParagraph": {
            "type": [
              "string",
              "null"
            ]
          },
          "links": {
            "type": "object",
            "required": [
              "aihot"
            ],
            "properties": {
              "aihot": {
                "type": "string",
                "format": "uri"
              }
            }
          },
          "attribution": {
            "$ref": "#/components/schemas/Attribution"
          }
        }
      },
...
    }
  }
}